Get rid of compilation warnings stemming from the fact that streamsize and size_t are not necessarily the same (at least on 32-bit build in MSVC 2010 streamsize is 64-bit when size_t is 32-bit and on 64-bit build in MSVC 2010 streamsize is signed when size_t is unsigned).
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@53326 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/test/pbacktest.cpp
@@ -158,7 +158,7 @@ static int s_StreamPushback(iostream&   ios,
         switch (j) {
         case 0:
             ios.read(data + nread, i);
-            j = ios.gcount();
+            j = (size_t)ios.gcount();
             _ASSERT(nread + j <= size);
             if (!ios.good()) {
                 _ASSERT(ios.rdstate() & NcbiEofbit);
@@ -172,7 +172,7 @@ static int s_StreamPushback(iostream&   ios,
             break;
         case 1:
             ios.get(data + nread, i + 1/*EOL*/, eol[0]);
-            j = ios.gcount();
+            j = (size_t)ios.gcount();
             _ASSERT(!data[nread + j]);
             if (nread + j < size)
                 data[nread + j] = orig[nread + j]; // undo '\0' side effect
@@ -198,7 +198,7 @@ static int s_StreamPushback(iostream&   ios,
             }
             break;
         default:
-            j = CStreamUtils::Readsome(ios, data + nread, i);
+            j = (size_t)CStreamUtils::Readsome(ios, data + nread, i);
             _ASSERT(nread + j <= size);
             break;
         }
