+++ src/corelib/test/test_ncbidiag_f_mt.cpp
@@ -462,7 +462,7 @@ bool CTestDiagApp::TestApp_Exit(void)
     TStrings messages;
 
     // Get the list of messages
-    CTempString ts(s_Sout.str(), s_Sout.pcount());
+    CTempString ts(s_Sout.str(), (size_t)s_Sout.pcount());
     NStr::Split(ts, "\r\n", messages);
     s_Sout.freeze(false);
